[Sync] Do not reupload entities for full-update types on reencryption
This CL makes it sure that entities are never reuploaded for full-update
data types as they are download-only. Ideally, such data types must
never be encrypted but it still wouldn't guarantee that data type state
persisted on the disk is correct (e.g. it could contain an encryption
key name even when the data type is non-encryptable).

Bug: 455150916
Change-Id: Ibc16437b0815849616562dfccdaf8352fc8c9286
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7131509
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1542475}