Move NativeWindowId to a new capture_util.h header
This is in preparation for the introduction of a new
PipScreenCaptureCoordinatorProxy class in a follow-up CL which will
depend on having the same definition of this type as
PipScreenCaptureCoordinator and GetNativeWindowIdMac().

The initial definition of NativeWindowId is int64_t to represent
window numbers on macOS, but this can later be adapted to other
platforms through ifdefs in capture_util.h.

Bug: 445202459
Change-Id: I15fc9c5bbeac6ef62a2ec18d693f315675ad6275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7132018
Reviewed-by: Johannes Kron <kron@chromium.org>
Commit-Queue: Tove Petersson <tovep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542510}