[FedCM] Convert RequestService to DocumentUserData
Currently, RequestService is a DocumentService. While this
provides some convenient mojo helpers,  it does not guarantee
that there is only one RequestService for a document, and
there is no way to retrieve the current RequestService for
a document if one exists.

Because the renderer caches the mojo connection, this has not
been an issue in practice, but we now would like to initiate
a FedCM request from the browser process, and so it will
be very useful to be able to retrieve/create a unique
RequestService for the current document, so this CL makes
that change.

See also previous discussion here:
https://chromium.slack.com/archives/CGJNSDHA7/p1721248784924949

Bug: 455800279
Change-Id: I4e8b34a1ba0f0f19a85ea6dfc1671a8591a338cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7115160
Commit-Queue: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542614}