[Partitioned Popins Removal] Remove Counters
This is the first of a series of CLs to remove the popins code.
The Popins code touches a lot of sensitive areas so needs real review.
This is being done in multiple steps to reduce reviewer overhead.
The removal will be completed before the M144 branch so risk is low.
This only ever got to the DevTrial phase so no formal deprecation or
removal process is needed.

We're starting with the counters because it's an easy review, but the
review bar is high as it touches the UKM file.

Bug: 458633862
Change-Id: I91e5a4e991674853c4ced35f67053f673ee96549
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7135463
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sun Yueru <yrsun@chromium.org>
Commit-Queue: Sun Yueru <yrsun@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542645}