[Bookmarks Bar] Fix blank space on NTP
This CL fixes the bug of an empty space appearing when the bookmark bar
is turned on in the new tab page. The issue is caused by the offset
of the Toolbar being incorrect and not offset by the bookmark bar
height. The offset is not correct because we are calculating the offset
before we set the height supplier for the bookmark bar. The second time
the bookmark bar is turned on the bug does not appear because the height
supplier has already been set. We update the TabbedRootUiCoordinator
construction of the BookmarkBarCoordinator to wait until after the
height supplier has been set on the ToolbarManager before requesting a
layer update.

Bug: 458594867
Change-Id: Ic1f76b0f87fea5409448d9cb956a5870465a8488
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7137243
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542653}