Ensure navigations aborted by tab closure report a net error.
When a tab is closed mid-navigation, the navigation is not marked as
aborted. This causes a race in our UrlOverridingTests with tests that
close the tab upon redirecting to an app.
Since we allow navigations to proceed in parallel with the external
navigation checks, the way navigations typically get aborted is when
the response to a redirect comes in, and the URL that redirected
launched an app, we abort that response (either synchronously if the
external navigation checks completed, or we defer until the navigation
checks on the redirect complete and then abort). However, if the tab
is closed before the response to the redirect comes in, the navigation
is canceled by the tab closure. Since our tests want to ensure the
navigation was correctly aborted rather than committing, the tab
closure should also mark the navigation as aborted.

With this change, a tab closure/WebContents destruction mid-navigation
will cause the DidFinishNavigation observer call's NavigationHandle
to be marked as aborted (instead of finished without error).

For Android TabObservers, this change allows didFailNavigation to be
delivered after onDestroyed events if the tab destruction causes a
navigation to fail.

This change is guarded by the kAbortNavigationsFromTabClosures kill
switch.

Bug: 446837985
Change-Id: I51a3848b43757793302e696448d5e6e56a8140a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990009
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542690}