Support fallback to well-known providers in DoH Automatic Mode.
This CL gives the DnsClient the ability to use a fallback list of DoH
nameservers when Secure DNS has fails to upgrade the DNS provider to
DoH in automatic mode.

This logic is currently dormant as the feature is disabled and no
embedder populates the fallback list yet.

Further context for this change:

Secure DNS uses DNS-over-HTTPS (DoH) to encrypt users DNS queries. In
'automatic' mode, it reads the system DNS provider then attempts to
upgrade it to DoH by comparing with a hardcoded list of providers. If
Chrome can upgrade it to DoH, it'll use DoH for DNS queries and fall
back to insecure DNS if DoH fails.

Secure DNS is enabled by default in automatic mode, but many clients
are falling back to insecure DNS due to their DNS provider not being in
Chrome's list. This new Secure DNS behavior can be enabled and used by
users who want to use DoH even if their provider doesn't support it.

Bug: 448683318
Change-Id: I0351619dd695e713f853d3ab594924cd02dca982
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7042523
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Kirubel Aklilu <kaklilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542750}