Add a pref to enable Secure DNS 'automatic mode with DoH fallback'.
This CL plumbs a vector<IPEndpoint> of fallback DoH nameservers from
the browser process to the network service to the network stack. When
the feature is enabled this list contains the Google DNS address and
is empty otherwise.

These DoH providers will only be used in Secure DNS automatic mode if
the pref is enabled, this feature is enabled, and automatic mode has
already failed to upgrade the existing DNS provider to DoH.

Note: this fallback is not yet supported on Chrome OS.

Bug: 448683318
Change-Id: I1739fdce850b1bb4bc9c7376df23750c32313e3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128138
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Kirubel Aklilu <kaklilu@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Andrew Williams <awillia@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542754}