Fix selected icon color and adds RTL to bookmark bar folder icon
Increase contrast makes the bookmark bar menu folder icon disappear when
hovered over. This is broken because the MenuItemView was unable to
change the icon color on selection.

Fix this issue by changing the icon image to a VectorIconModel. This
allows the MenuItemView to change the icon color on selection.

Because we're changing the image to a VectorIconModel, we also need to
add RTL to the icon itself.

Bug: 459502974
Change-Id: I9947c4784c32d91fc3a6af73b5db487f8fe4a5c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7137315
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Commit-Queue: David Yeung <dayeung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542778}