Add PDF.InfoBar.SettingsResult histogram
This change adds a histogram where the PDF infobar (which offers to set
Chrome as the default PDF viewer) is shown. This histogram is intended
to give insight into which settings UI is shown, and whether users
successfully set Chrome as the default. It does so by checking whether
Chrome is the default PDF viewer 30 seconds after the settings UI was
shown.

This change also updates
`ShellUtil::ShowSetDefaultForFileExtensionSystemUI()` to return which
settings UI was shown.

NO_IFTTT=Adding if-this-then-that block for the first time.

Bug: 396202897
Change-Id: I178757ff958888ff784352b1dc4292949928c42e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7126095
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542803}