[bedrock] Migrate browser_conditions.cc away from BrowserListObserver.
Note that this changes the existing observer from listening to all
browsers globally to just listening to browsers within the relevant
Profile.

This migration is part of project bedrock to reduce the dependencies on
Browser and BrowserList. See https://crbug.com/431671320 for more info.

Bug: 459505408
Change-Id: Ideecccd6e8e70803468a74eec014c50c6a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088444
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Glenn Hartmann <hartmanng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542861}