extensions: Port downloads API tests to desktop Android, part 11
Some tests rely on file paths, but Android uses content URIs for its
downloads. These URIs cannot be reliably predicted in advance (you can
try, but your tests will be flaky). Therefore we can't use the current
approach of waiting for expected events with the filename specified.

Instead, check each event router event for a filename, and if one is
present, cache it. For Android, at the end of the event stream (which
events with a filename) get the last cached filename, and make sure
it's a valid URI. I talked to the downloads team and just checking
the URI format was their suggested approach (which makes sense to me
since I can't reliably predict the whole URI).

Bug: 405219117
Change-Id: I95da97585880e54d899ce1e2cfa23cbb2c8b9f43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7138840
Reviewed-by: Michael Wojcicka <mwoj@google.com>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542869}