extensions: Port downloads API tests to desktop Android, part 12
Now that we've decided we're using content URIs on Android, many of
the tests we're currently skipping can be mechanically enabled.

Basically, they need to skip the event check for an update with a
filename (since they don't know the filename on Android) and instead
check the filename at the end of the test.

Do that to a bunch of tests.

Bug: 405219117
Change-Id: Ibcd98895963b7ffba0a40a9d1bd4535aec22ca47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7139578
Reviewed-by: Michael Wojcicka <mwoj@google.com>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542870}