[bedrock] Migrate session_service_base.cc away from BrowserListObserver.
Note that this changes the existing observer from listening to all
browsers globally to just listening to browsers within the relevant
Profile.

This migration is part of project bedrock to reduce the dependencies on
Browser and BrowserList. See https://crbug.com/431671320 for more info.

Bug: 459505408
Change-Id: Ifb47104029697a378ced847b14a884736a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7091712
Commit-Queue: Glenn Hartmann <hartmanng@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542884}