[glic] Round the glic window on ChromeOS
For ChromeOS, it's essential to explicitly round the contents of the
ClientView. The FrameViewAsh manages the rounding of the header (which
isn't painted in the case of a glic window) and then delegates the
rounding of the ClientView's content.

Window decorations, including the shadow, resize shadow, and highlight
border, automatically adapt to the specified window radius set by
`Widget::InitParams::rounded_corners`.

Bug: b:452137970
Change-Id: I42ae5b2d19f5a7dd248efd6e3b578915d47e0c5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7138321
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542898}