Move tab observing logic out of MediaCapturePickerDialog
This CL moves the (currently very little) tab observing logic out of
MediaCapturePickerDialog. This is so the logic can be shared with the
new picker. In the future we will need more logic here to deal with
policy that filters what tabs are shown in the picker.

Bug: 352187279
Test: CQ + unit test
Change-Id: I2249edc11e5739d31809306f626715587f4fd7e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7137018
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542912}