IDB: remove custom container class `list_set`
This container has no apparent purpose any more (presumably, it used to
be more helpful). Under the hood, it maintains both a list and a set so
that existence checks can be fast while order is also maintained.
However, the only place that code uses the existence check is during
insertion (implicitly) and in a DCHECK (right after insertion, which was
overkill). Neither of these existence checks is necessary.

Bug: none
Change-Id: I389136997675280d6e8c1334163fd817d7eb9f96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7135162
Reviewed-by: Gabriel Brito <gabrielbrito@microsoft.com>
Commit-Queue: Evan Stade <evanstade@microsoft.com>
Reviewed-by: Evan Stade <evanstade@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1542924}