Glic actor: Don't wait for visual update from non-captured background tab
When creating a new tab, the tab is not associated with an actor task
until after the creation tool is complete. If the tab is created in the
background, it won't get visual updates and we delay unnecessarily
before timing out and continuing to the next navigation action.

Bug: b:454961534, b:459594544
Change-Id: I4df4628caf624d97353a3d1bc9dad51abf92fb00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7139185
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542929}