extensions: Port downloads API tests to desktop Android, part 13
Continue to fix tests that rely on knowing the filename in advance.
This fixes many of the currently skipped tests. What's left are tests
that depend on dialogs we haven't ported yet, some networking details,
and downloads UI that doesn't exist on Android (like the old shelf).

Bug: 405219117
Change-Id: Ic7447739603d944354ac20412e871f0c24bf25b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7138359
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Michael Wojcicka <mwoj@google.com>
Cr-Commit-Position: refs/heads/main@{#1542955}