Follow-up crash fix in AnchorElementInteractionTracker::ViewportIntersectionUpdate()
This CL is a follow-up of a crash bug fix in crrev.com/c/7129521.

The previous CL fixes crashes when `anchor`'s URL is empty when an
element is entered in viewport. This CL handles `anchor` in the same
way also when the element left viewport. Note that the case in this CL
is relatively minor to the original case, and all changes are behind
the `PreloadingEagerViewportHeuristics` flag.

Bug: 436705485, 458237344
Change-Id: If896e047aa4919d920ad2e4c30093f23162c43df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7140201
Commit-Queue: Takashi Nakayama <tnak@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542994}