Add LINT.IfChange/ThenChange for PasswordSpecificsData trimming
Note that there are some deprecated fields which are not trimmed. It's
better to keep it as is, otherwise the metadata would be redownloaded if
those fields started to be cleared in this CL.

The message Notes had to be moved because nested Lint.IfChanges are not
supported.

NO_IFTTT=These are new IFTTT

Bug: 454561071
Change-Id: I6e871381ef4fade393e33d8b4af8efc87e35dc8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7131829
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1543018}