[Autofill] Turn ExtractForm() into ExtractFormWithField()
AutofillDriver::ExtractForm() takes a *renderer* form's FormGlobalId
and extracts the respective browser form.

The mix of renderer and browser forms is confusing and violates the
principle that renderer forms should be encapsulated by AutofillDriver
and AutofillDriverRouter.

This CL changes the function to take a FieldGlobalId instead. This
simplifies the contract because unlike with forms, there is no
distinction between renderer and browser fields.

For the only existing callsite of the function, this change is even
a simplification.

Bug: 40284824
Change-Id: I34eee9d7a4a8be4e31d20d367061867bc5497bd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7130218
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1543029}