Add BrowserAutofillManager::UpdateCache()
This should simplify the upcoming Finch experiment that will modify this
function so that the feature flag acts on a single method.

Details of the experiment can be found in
go/optimizing-autofill-cache-updates. The short version is that it will
aim at getting rid of FormStructure::RetrieveFromCache() by updating the
cached form in place.

Bug: 456719060
Change-Id: I52c8da57fac1760c816cca11af674f26c15218a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7132038
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1543069}