NavigationController: Fix view-source prefix loss on error pages and consecutive reloads
Currently, the view-source status of a navigation is tracked via a
view-source: VirtualURL on the NavigationEntry. However, navigations
that result in an error page discard the pending entry and create a new
one when the error page commits. This new entry does not preserve the
view-source: prefix in its VirtualURL(). This CL ensures that the new
NavigationEntry created for an error page in a view-source navigation
correctly sets the Virtual URL with the view-source: prefix. Demo before
fix:
https://screencast.googleplex.com/cast/NjA2MjczNjg1MjEyMzY0OHw1ZDVkNmJhYy1lOA
After fix:
https://screencast.googleplex.com/cast/NTMwOTkyMjA0NzY4ODcwNHw4NzQzMzM5Yy05Ng

Bug: 437913384
Change-Id: I29d362e9aafb339e7270d8756f1ffd7f23d7d5ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7062130
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Aliaksei Bahdzevich <thelex@google.com>
Cr-Commit-Position: refs/heads/main@{#1543153}