android-a11y: deprecate usage of ANI.CollectionItemInfo.setHeading
In API 28, AccessibilityNodeInfo.set|isHeading was introduced[1]. This
"moves" the idea of a heading from being specific to a
CollectionItemInfo to be generalized to any AccessibilityNodeInfo (with
or without a collection).

This also resulted in deprecation of AccessibilityNodeInfo.CollectionItemInfo[Compat].isHeading.[2][4] in API 28. It didn't result in the deprecation of the "heading" based constructor for CollectionItemInfo[3].

In Chrome, we adjust our implementation to simply expose the more
generic AccessibilityNodeInfo.setHeading. Within Android frameowrk code,
this is works in prior cases for CollectionItemInfo as well as the new
usage on ANI only.

1. https://developer.android.com/reference/android/view/accessibility/AccessibilityNodeInfo#isHeading()
2. https://developer.android.com/reference/android/view/accessibility/AccessibilityNodeInfo.CollectionItemInfo#isHeading()
3. https://developer.android.com/reference/android/view/accessibility/AccessibilityNodeInfo.CollectionItemInfo#CollectionItemInfo(int,%20int,%20int,%20int,%20boolean)
4. https://cs.android.com/search?q=CollectionItemInfoCompat%20%20isheading%20deprecated

Bug: 443079218

Change-Id: Ie290a5d4f40eb3ad80f6f6a8b1621829a7982fc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7138959
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1543169}